# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "erdsea-api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "paths": {
...
        "/transactions/address/{address}/{offset}/{limit}": {
            "get": {
                "description": "Retrieves transactions for an address. Unordered.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Gets transaction for an address.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/transactions/asset/{tokenId}/{nonce}/{offset}/{limit}": {
            "get": {
                "description": "Retrieves transactions for an asset. Unordered.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Gets transaction for an asset.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "nonce",
                        "name": "nonce",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/transactions/list/{offset}/{limit}": {
            "get": {
                "description": "Retrieves transactions. Unordered.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Gets transaction list.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/tx-template/buy-nft/{userAddress}/{tokenId}/{nonce}/{price}": {
            "get": {
                "description": "Retrieves tx-template for NFT buy. Only account nonce and signature must be added afterwards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tx-template"
                ],
                "summary": "Gets tx-template for NFT buy.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user address",
                        "name": "userAddress",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "nonce",
                        "name": "nonce",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "price",
                        "name": "price",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/formatter.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/tx-template/list-nft/{userAddress}/{tokenId}/{nonce}/{price}": {
            "get": {
                "description": "Retrieves tx-template for NFT list. Only account nonce and signature must be added afterwards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tx-template"
                ],
                "summary": "Gets tx-template for NFT list.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user address",
                        "name": "userAddress",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "nonce",
                        "name": "nonce",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "price",
                        "name": "price",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/formatter.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        },
        "/tx-template/withdraw-nft/{userAddress}/{tokenId}/{nonce}": {
            "get": {
                "description": "Retrieves tx-template for NFT withdraw. Only account nonce and signature must be added afterwards.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tx-template"
                ],
                "summary": "Gets tx-template for NFT withdraw.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user address",
                        "name": "userAddress",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "nonce",
                        "name": "nonce",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/formatter.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "data.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "data.Transaction": {
            "type": "object",
            "properties": {
                "assetID": {
                    "type": "integer"
                },
                "buyerID": {
                    "type": "integer"
                },
                "collectionID": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "priceNominal": {
                    "type": "number"
                },
                "sellerID": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "formatter.Transaction": {
            "type": "object",
            "properties": {
                "chainID": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "gasLimit": {
                    "type": "integer"
                },
                "gasPrice": {
                    "type": "integer"
                },
                "nonce": {
                    "type": "integer"
                },
                "options": {
                    "type": "integer"
                },
                "receiver": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    }
}